// .kilo/worktrees/rumbling-cello/frontend/src/types/inventory.ts
export interface InvGudang extends MasterData {
    penanggung_jawab_id?: number | null;
    department_id?: number | null;
    lokasi_kerja_id?: number | null;
    lokasi?: string;
    penanggung_jawab?: { id: number; nama_lengkap: string };
    department?: { id: number; nama: string };
    lokasi_kerja?: { id: number; nama: string; kode_site?: string };
}